docs: reformatting Error and Warning tables in RULES.md  (#318)
## Summary

Updating table structure for `Errors` and `Warnings` for clarity purposes. This added a `Cause` section for the HIPAA rules and separated out  Control IDs in `NIST 800-53`. I believe this makes it easier for users to lookup information about implemented rules.

No updates to `Excluded Rules` tables

### New 
<html><body>
<!--StartFragment-->

| Rule ID | Cause | Explanation |
| -- | -- | -- | 

or if the Pack has control IDs 

| Rule ID | Cause | Explanation | Relevant Control ID(s) |
| -- | -- | -- | -- |

<!--EndFragment-->
</body>
</html>

###  Old

<html><body>
<!--StartFragment-->

| Rule ID | Summary | Explanation
| -- | -- | --


<!--EndFragment-->
</body>
</html>




----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*