feat(core): simpler rule creation system with error handling (#391)
Resolves #390 
Related to #388

An initial implementation of a new rule validation system. However there are still some cases where token values that resolve to Intrinsic functions can pass validation of rules due to. Take the following example

```typescript
    const stack = new Stack();
    Aspects.of(stack).add(new AwsSolutionsChecks());
    const pDbPort = new CfnParameter(stack, 'pDbPort', {
      type: 'Number',
      default: 27017,
      description: 'The port number to use for the DocumentDB'
    });
    new CfnDBCluster(stack, 'rDocDB', {
      masterUsername: 'foo',
      masterUserPassword: 'bar',
      port: pDbPort.valueAsNumber
    });
  ``` 
  It will pass the [AwsSolutions-DOC2 rule](https://github.com/cdklabs/cdk-nag/blob/46af546307a871e0750c416e10c7dbd4f64c1a86/src/AwsSolutions/rules/databases/DOC2.ts) since the token resolves to a encoded number. 

Additionally even  if the token is resolved, it still resolves to a `{Ref: 'pDbPort'}` which passes the equality test in the rule. This potentially deserves an issue of it's own as I'm not sure if they is an easy way to fix this issue.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*