refactor(nag-pack): separate NagPack.applyRule functionality to improve extensibility
Currently `applyRule` handles both performing validation checks and applying annotation
regarding the results in a single method. The way in which non-compliance (Error|Warning)
is recorded is through the `Annations.of(node).addError|Warning()` which utilizes the
native CDK Error+Warning handling and is considered "blocking" during synth. Generally
having non-compliant rules result in blocked synth is ideal, but during rapid development
it would be prefered to just get reports of the errors while stilling being able to synth.

Additionally, being able to extend the reporting functionality of nag packs without
having to rewrite the core validation logic would be preferred and is more maintainable.

This refactor does not change any functionality of the `applyRule` functionality, it
simply splits it up into smaller methods that can be extended/overwritten by NagPacks
to support additional use cases.