feat(nag-logger): provide loggers with original rule name (#1550)
For use cases where we want to build further automation around CDK Nag, it's useful to be able to identify the original CDK Nag rule which was applied. This change provides loggers with the original rule name.

The practical use case I'm looking at is automating the generation of base [Threat Composer](https://github.com/awslabs/threat-composer) threat models. The idea is that based on CDK Nag's introspection of your infrastructure, we can map CDK Nag rules to particular threats that they would mitigate, and so provide a good starting point from which to build a threat model for an application.

With the rule name available in a logger, I only need to worry about mapping CDK Nag rule names to threats, rather than maintaining mappings for each nag pack's rule ID, or only supporting a specific nag pack :)